BackEnd/Agent/requirements.txt
langgraph
langchain-google-genai
langchain-core
langchain-community
fastapi
uvicorn
python-dotenv
pydantic
httpx
youtube-transcript-api
sqlmodel
psycopg2-binary
langgraph-checkpoint-postgres